% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsIOData.R
\name{ivsIOData}
\alias{ivsIOData}
\title{Perform Input Variable Selection}
\usage{
ivsIOData(y, x, ivsm, ivs_param)
}
\arguments{
\item{y}{target vector [N x 1]}

\item{x}{input matrix [N x D]}

\item{ivsm}{input variable selection (IVS) method [string:'none','boruta','rrf','ea_cmi_htc','ea_cmi_tol','knn_cmi_tol','knn_cmi_bi_tol',pmis_bic','pcis_bic']}

\item{ivs_param}{parameters for input variable selection method
\itemize{
\item rrf
\itemize{
\item ivs_param[1] = regularization parameter [0 < scalar <= 1]
\item ivs_param[2] = ntrees [integer > 1)] usually between 128 and 1000
}
\item ea_cmi
\itemize{
\item ivs_param[1] =  Hampel test criterion threshold [scalar > 0]
}
\item ea_cmi_tol
\itemize{
\item ivs_param[1] = CMI/MI threshold [0 < scalar <= 1]
}
\item knn_cmi_tol
\itemize{
\item ivs_param[1] = CMI/MI threshold [0 < scalar <= 1]
\item ivs_param[2] = no. of nearest neighbours [1 < integer < sample size-1]
}
}}
}
\value{
list that contains indices and names of selected input variables
}
\description{
This function performs input variable selection for a given input-output
dataset.
}
\references{
Quilty, J., J. Adamowski, B. Khalil, and M. Rathinasamy (2016), Bootstrap rank-
ordered conditional mutual information (broCMI): A nonlinear input variable
selection method for water resources modeling, Water Resour. Res., 52,
doi:10.1002/2015WR016959.

Quilty, J., and J. Adamowski (2018), Addressing the incorrect usage of wavelet-
based hydrological and water resources forecasting models for real-world
applications with best practices and a new forecasting framework, J. Hydrol.,
doi:10.1016/j.jhydrol.2018.05.003.
}
\author{
John Quilty
}
