% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMI_PCIS_bic.R
\name{PCIS_framework_bic}
\alias{PCIS_framework_bic}
\title{Partial Correlation Input Selection (PCIS)}
\usage{
PCIS_framework_bic(x, y, silent = FALSE)
}
\arguments{
\item{x}{A matrix of input variables as columns and observations as rows [N x D]}

\item{y}{A vector containing target values}

\item{silent}{Verbosity, Default: \code{FALSE}}
}
\value{
A data frame with selected inputs (integer indices) in the first column, and additional information in the other columns.
}
\description{
An implementation of the partial correlation input selection algorithm (May et al., 2008)
with the Bayesian Information Criterion (BIC) used for the stopping criterion. This function is essentially the same as
\code{PMI_grnn_framework_bic}, except that the Pearson correlation coefficient and multiple linear regression are used instead of
partial mutual information (PMI) and the generalied regresssion neural network (GRNN), respectively.

This function was written by Greer B. Humphrey (School of Civil, Environmental, and Mining Engineering,
University of Adelaide, SA, 5005, Australia). The original code repository for this function
can be found at (https://github.com/gbhumphrey1/PMIS_PCIS).
}
\examples{
\dontrun{
if(interactive()){
 X <- matrix(rnorm(10*1000), ncol = 10)
 y <- rnorm(1000)
 res <- PCIS_framework_bic(X, y)
 res$Input # selected inputs
 }
}
}
\references{
Galelli, S., Humphrey, G. B., Maier, H. R., Castelletti, A., Dandy, G. C., & Gibbs, M. S. (2014).
An evaluation framework for input variable selection algorithms for environmental data-driven models.
Environmental Modelling & Software, 62, 33–51. https://doi.org/10.1016/j.envsoft.2014.08.015

May, R. J., Maier, H. R., Dandy, G. C., & Fernando, T. M. K. G. (2008). Non-linear variable selection for
artificial neural networks using partial mutual information. Environmental Modelling & Software, 23(10-11),
1312–1326. https://doi.org/10.1016/j.envsoft.2008.03.007
}
\seealso{
\code{\link[stats]{lm.fit}}, \code{\link[stats]{cor}}
}
