% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMI_PCIS_bic.R
\name{PMI_grnn_framework_bic}
\alias{PMI_grnn_framework_bic}
\title{Partial Mutual Information Input Selection using the Generalized Regression Neural Network}
\usage{
PMI_grnn_framework_bic(x, y, ybin = FALSE, silent = FALSE)
}
\arguments{
\item{x}{A matrix of input variables as columns and observations as rows [N x D]}

\item{y}{A vector containing target values}

\item{ybin}{Is \code{y} binary?, Default: FALSE}

\item{silent}{Verbosity, Default: \code{FALSE}}
}
\value{
A data frame with selected inputs (integer indices) in the first column, and additional information in the other columns.
}
\description{
From Galelli S., Humphrey G.B., Maier H.R., Castelletti A., Dandy G.C. and Gibbs M.S. (2014) <DOI: 10.1016/j.envsoft.2014.08.015>:

The PMIS algorithm adopts a forward selection strategy (i.e. one variable is selected at each iteration) based on the
estimation of the PMI, which measures the partial dependence between each input variable and the output, conditional
on the inputs that have already been selected. To estimate the PMI, the algorithm adopts a kernel density estimator,
whose accuracy depends on the value of a smoothing parameter (or bandwidth) lambda. Similarly to Sharma (2000) and
Bowden et al. (2005a), the Gaussian reference bandwidth (Scott, 1992) is adopted, because of its simplicity and computational efficiency.
The calculation of PMI also requires estimation of residual information in the input variables once the effect of the already
selected inputs has been taken in consideration: this is done through the identification of a General Regression Neural Network (GRNN),
which is a nonlinear and nonparametric regression method (Li et al., 2014). In addition to lambda, the other parameter to be set is the
stopping criterion, which is based on the coefficient of determination R2: the PMIS algorithm is stopped when the selection of a further
input variable leads to a decrease of R2 in the underlying GRNN being identified.

This function was written by Greer B. Humphrey (School of Civil, Environmental, and Mining Engineering,
University of Adelaide, SA, 5005, Australia). The original code repository for this function
can be found at (https://github.com/gbhumphrey1/PMIS_PCIS).
}
\examples{
\dontrun{
if(interactive()){
 X <- matrix(rnorm(10*1000), ncol = 10)
 y <- rnorm(1000)
 res <- PMI_grnn_framework_bic(X, y)
 res$Input # selected inputs
 }
}
}
\references{
Galelli, S., Humphrey, G. B., Maier, H. R., Castelletti, A., Dandy, G. C., & Gibbs, M. S. (2014).
An evaluation framework for input variable selection algorithms for environmental data-driven models.
Environmental Modelling & Software, 62, 33–51. https://doi.org/10.1016/j.envsoft.2014.08.015

May, R. J., Maier, H. R., Dandy, G. C., & Fernando, T. M. K. G. (2008). Non-linear variable selection for
artificial neural networks using partial mutual information. Environmental Modelling & Software, 23(10-11),
1312–1326. https://doi.org/10.1016/j.envsoft.2008.03.007

Sharma, A. (2000). Seasonal to interannual rainfall probabilistic forecasts for improved water supply management:
Part 1 — a strategy for system predictor identification. Journal of Hydrology, 239(1-4), 232–239.
https://doi.org/10.1016/s0022-1694(00)00346-2

Bowden, G. J., Dandy, G. C., & Maier, H. R. (2005). Input determination for neural network models in water resources applications.
part 1—background and methodology. Journal of Hydrology, 301(1-4), 75–92. https://doi.org/10.1016/j.jhydrol.2004.06.021

Scott, D. W. (2015). Multivariate density estimation: Theory, practice, and visualization. John Wiley & Sons.

Li, X., Zecchin, A. C., & Maier, H. R. (2014). Selection of smoothing parameter estimators for general regression neural
networks – applications to hydrological and water resources modelling. Environmental Modelling & Software, 59, 162–186.
https://doi.org/10.1016/j.envsoft.2014.05.010
}
\seealso{
\code{\link[stats]{nls}}, \code{\link[stats]{predict}}
\code{\link[utils]{write.table}}
}
